requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
click>=8.1.0
rich>=13.6.0
flask>=2.3.0
gunicorn>=21.2.0
